# Crypto: use ED25519 instead of RSA for better performance