# Description

This PR changes the asymmetric cryptography algorithm from RSA to ED25519, to reduce the peering time (which includes a new key generation), as well as to uniform key generation operations across the repo.

## Benchmarks
```
RSA 2048 key generation took on average 296.144ms
RSA 4096 key generation took on average 3190.145ms
ECDSA P224 key generation took on average 1.179ms
ECDSA P256 key generation took on average 1.017ms
ECDSA P384 key generation took on average 5.250ms
ECDSA P521 key generation took on average 8.622ms
ED25519 key generation took on average 0.082ms

RSA 2048 signature took on average 2.3094ms
ECDSA P256 signature took on average 0.0778ms
ED25519 signature took on average 0.1235ms
```

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Please also note any relevant details for your test configuration.

- [X] Unit testing (existing)
- [X] E2E testing (existing)